+++ pkg/analyzer/CHANGELOG.md
@@ -6,6 +6,8 @@
   classes (such as `RecursiveAstVisitor`) so that clients that extend those
   other classes will not be impacted.
 * Removed `EMPTY_LIST` constants.  Please use `const <...>[]` instead.
+* Disabled support for the task model.  Please use the new `AnalysisSession`
+  API.
 
 ## 0.34.2
 * Removed support for the `@checked` annotation.  Please use the `covariant`